requirements.txt
# Crossplane Automation System Dependencies

# Core dependencies
pyyaml>=6.0
requests>=2.28.0
pathlib2>=2.3.7

# OpenAI for LLM integration
openai>=1.0.0

# GitHub API integration
pygithub>=1.59.0

# Optional: Enhanced CLI experience
rich>=13.0.0
click>=8.1.0

# Development and testing
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Type checking
mypy>=0.900

# Documentation (optional)
mkdocs>=1.5.0
mkdocs-material>=9.0.0
